release-wes001 Corrigido models, reposity e integração com bd

+++ src/main/java/com/duodev/duodevbackend/repository/CompetenciaMentoriaRepository.java
@@ -3,5 +3,5 @@ package com.duodev.duodevbackend.repository;
 import com.duodev.duodevbackend.model.CompetenciaMentoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompetenciaMentoriaRepository extends JpaRepository<CompetenciaMentoria, Long>{
+public interface CompetenciaMentoriaRepository extends JpaRepository<CompetenciaMentoria, Integer>{
 }
